+++ signxml/processor.py
@@ -1,5 +1,7 @@
+import importlib.resources
 import logging
-import os
+import threading
+from functools import lru_cache
 from typing import Any, List, Tuple
 from xml.etree import ElementTree as stdlibElementTree
 
@@ -14,18 +16,26 @@ from .util import namespaces
 logger = logging.getLogger(__name__)
 
 
+@lru_cache
+def get_schema(schema_file: str) -> etree.XMLSchema:
+    pkg_name = "signxml.xades.schemas" if schema_file.startswith("XAdES") else "signxml.schemas"
+
+    pkg_resources = importlib.resources.files(pkg_name)
+    with pkg_resources.joinpath(schema_file).open("r") as schema_fh:
+        return etree.XMLSchema(etree.parse(schema_fh))
+
+
 class XMLProcessor:
     _schemas: List[Any] = []
     schema_files: List[Any] = []
     _default_parser, _parser = None, None
-    _schema_dir = os.path.normpath(os.path.join(os.path.dirname(__file__), "schemas"))
 
     @classmethod
     def schemas(cls):
-        if len(cls._schemas) == 0:
-            for schema_file in cls.schema_files:
-                schema_path = os.path.join(cls._schema_dir, schema_file)
-                cls._schemas.append(etree.XMLSchema(etree.parse(schema_path)))
+        with threading.Lock():
+            if len(cls._schemas) == 0:
+                for schema_file in cls.schema_files:
+                    cls._schemas.append(get_schema(schema_file))
         return cls._schemas
 
     @property
@@ -38,6 +48,9 @@ class XMLProcessor:
 
     def _fromstring(self, xml_string, **kwargs):
         xml_node = etree.fromstring(xml_string, parser=self.parser, **kwargs)
+        docinfo = etree.ElementTree(xml_node).docinfo
+        if docinfo.internalDTD is not None or docinfo.externalDTD is not None:
+            raise InvalidInput("DTD declarations are not supported in XML input")
         for entity in xml_node.iter(etree.Entity):
             raise InvalidInput("Entities are not supported in XML input")
         return xml_node
@@ -62,19 +75,22 @@ class XMLSignatureProcessor(XMLProcessor):
     schema_files = ["xmldsig1-schema.xsd"]
 
     # See https://tools.ietf.org/html/rfc5656
+    # ec.SECT* curves have been de-registered due to CVE-2026-26007
+    # (https://github.com/pyca/cryptography/security/advisories/GHSA-r6ph-v2qm-q3c2)
+    # List of disabled curves:
+    # "urn:oid:1.3.132.0.1": ec.SECT163K1,
+    # "urn:oid:1.3.132.0.26": ec.SECT233K1,
+    # "urn:oid:1.3.132.0.27": ec.SECT233R1,
+    # "urn:oid:1.3.132.0.16": ec.SECT283R1,
+    # "urn:oid:1.3.132.0.36": ec.SECT409K1,
+    # "urn:oid:1.3.132.0.37": ec.SECT409R1,
+    # "urn:oid:1.3.132.0.38": ec.SECT571K1
     known_ecdsa_curves = {
         "urn:oid:1.2.840.10045.3.1.7": ec.SECP256R1,
         "urn:oid:1.3.132.0.34": ec.SECP384R1,
         "urn:oid:1.3.132.0.35": ec.SECP521R1,
-        "urn:oid:1.3.132.0.1": ec.SECT163K1,
         "urn:oid:1.2.840.10045.3.1.1": ec.SECP192R1,
         "urn:oid:1.3.132.0.33": ec.SECP224R1,
-        "urn:oid:1.3.132.0.26": ec.SECT233K1,
-        "urn:oid:1.3.132.0.27": ec.SECT233R1,
-        "urn:oid:1.3.132.0.16": ec.SECT283R1,
-        "urn:oid:1.3.132.0.36": ec.SECT409K1,
-        "urn:oid:1.3.132.0.37": ec.SECT409R1,
-        "urn:oid:1.3.132.0.38": ec.SECT571K1,
     }
     known_ecdsa_curve_oids = {ec().name: oid for oid, ec in known_ecdsa_curves.items()}  # type: ignore[abstract]